// src/app/models/employee.model.ts
export class Employee {
  id?: number = 0;
  firstname: string = '';
  lastname: string = '';
  birthdate: string = '';
  gender: string = '';
  education: string = '';
  jobExperience: number = 0;
  profile: string = '';
  lastPromotion: string='';
  gid: string= '';
  doj: string = '';
  jobDesignation:string=''; 
  awardsReceived:string='';
  awardName:string='';
  awarddate: string='';
}